applications: connectivity_bridge: patch usb_serial_str desc
if CONFIG_SPEED_OPTIMIZATIONS=n, the compiler catches a buffer
overflow in usb_update_sn_string_descriptor() where the destination
usb_serial_str is too small for the full descriptor.

This commit updates the size of the usb_serial_str buffer to match
adapt to the size of HW_ID_LEN or the custom descriptor used for
the NRF52X SOC (which doesn't support hw_id_get()).

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>